refactor: mover funções de formatação para um novo módulo
- Atualizadas as taxas de cobrança por hora para CAR e MOTORCYCLE.
- Funções de formatação de moeda e duração foram extraídas para o novo arquivo utils/formatters.ts, melhorando a organização do código.
